moves error types of out `lib.rs
