+++ src/pointer.rs
@@ -1,6 +1,7 @@
-use crate::{InvalidEncodingError, ParseError, ReplaceTokenError, Token, Tokens};
+use crate::{InvalidEncodingError, Token, Tokens};
 use alloc::{
     borrow::ToOwned,
+    fmt,
     string::{String, ToString},
     vec::Vec,
 };
@@ -869,6 +870,180 @@ const fn validate(value: &str) -> Result<&str, ParseError> {
 ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
 ╔══════════════════════════════════════════════════════════════════════════════╗
 ║                                                                              ║
+║                                  ParseError                                  ║
+║                                 ¯¯¯¯¯¯¯¯¯¯¯¯                                 ║
+╚══════════════════════════════════════════════════════════════════════════════╝
+░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
+*/
+
+/// Indicates that a `Pointer` was malformed and unable to be parsed.
+#[derive(Debug, PartialEq)]
+pub enum ParseError {
+    /// `Pointer` did not start with a backslash (`'/'`).
+    NoLeadingBackslash,
+
+    /// `Pointer` contained invalid encoding (e.g. `~` not followed by `0` or
+    /// `1`).
+    InvalidEncoding {
+        /// Offset of the partial pointer starting with the token that contained
+        /// the invalid encoding
+        offset: usize,
+        /// The source `InvalidEncodingError`
+        source: InvalidEncodingError,
+    },
+}
+
+impl fmt::Display for ParseError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::NoLeadingBackslash { .. } => {
+                write!(
+                    f,
+                    "json pointer is malformed as it does not start with a backslash ('/')"
+                )
+            }
+            Self::InvalidEncoding { source, .. } => write!(f, "{source}"),
+        }
+    }
+}
+
+impl ParseError {
+    /// Returns `true` if this error is `NoLeadingBackslash`; otherwise returns
+    /// `false`.
+    pub fn is_no_leading_backslash(&self) -> bool {
+        matches!(self, Self::NoLeadingBackslash { .. })
+    }
+
+    /// Returns `true` if this error is `InvalidEncoding`; otherwise returns
+    /// `false`.
+    pub fn is_invalid_encoding(&self) -> bool {
+        matches!(self, Self::InvalidEncoding { .. })
+    }
+
+    /// Offset of the partial pointer starting with the token which caused the
+    /// error.
+    /// ```text
+    /// "/foo/invalid~tilde/invalid"
+    ///      ↑
+    ///      4
+    /// ```
+    /// ```
+    /// # use jsonptr::PointerBuf;
+    /// let err = PointerBuf::parse("/foo/invalid~tilde/invalid").unwrap_err();
+    /// assert_eq!(err.pointer_offset(), 4)
+    /// ```
+    pub fn pointer_offset(&self) -> usize {
+        match *self {
+            Self::NoLeadingBackslash { .. } => 0,
+            Self::InvalidEncoding { offset, .. } => offset,
+        }
+    }
+
+    /// Offset of the character index from within the first token of
+    /// [`Self::pointer_offset`])
+    /// ```text
+    /// "/foo/invalid~tilde/invalid"
+    ///              ↑
+    ///              8
+    /// ```
+    /// ```
+    /// # use jsonptr::PointerBuf;
+    /// let err = PointerBuf::parse("/foo/invalid~tilde/invalid").unwrap_err();
+    /// assert_eq!(err.source_offset(), 8)
+    /// ```
+    pub fn source_offset(&self) -> usize {
+        match self {
+            Self::NoLeadingBackslash { .. } => 0,
+            Self::InvalidEncoding { source, .. } => source.offset,
+        }
+    }
+
+    /// Offset of the first invalid encoding from within the pointer.
+    /// ```text
+    /// "/foo/invalid~tilde/invalid"
+    ///              ↑
+    ///             12
+    /// ```
+    /// ```
+    /// # use jsonptr::PointerBuf;
+    /// let err = PointerBuf::parse("/foo/invalid~tilde/invalid").unwrap_err();
+    /// assert_eq!(err.pointer_offset(), 4)
+    /// ```
+    pub fn complete_offset(&self) -> usize {
+        self.source_offset() + self.pointer_offset()
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for ParseError {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        match self {
+            Self::InvalidEncoding { source, .. } => Some(source),
+            Self::NoLeadingBackslash => None,
+        }
+    }
+}
+
+/*
+░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
+╔══════════════════════════════════════════════════════════════════════════════╗
+║                                                                              ║
+║                                NotFoundError                                 ║
+║                               ¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯                                ║
+╚══════════════════════════════════════════════════════════════════════════════╝
+░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
+*/
+
+/// An error that indicates a [`Pointer`]'s path was not found in the data.
+#[derive(Debug, PartialEq, Eq)]
+pub struct NotFoundError {
+    /// The starting offset of the [`Token`] within the [`Pointer`] which could not
+    /// be resolved.
+    pub offset: usize,
+}
+
+impl fmt::Display for NotFoundError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "path starting at offset {} not found", self.offset)
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for NotFoundError {}
+
+/*
+░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
+╔══════════════════════════════════════════════════════════════════════════════╗
+║                                                                              ║
+║                              ReplaceTokenError                               ║
+║                             ¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯¯                              ║
+╚══════════════════════════════════════════════════════════════════════════════╝
+░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
+*/
+
+/// Returned from `Pointer::replace_token` when the provided index is out of
+/// bounds.
+#[derive(Debug, PartialEq, Eq)]
+pub struct ReplaceTokenError {
+    /// The index of the token that was out of bounds.
+    pub index: usize,
+    /// The number of tokens in the `Pointer`.
+    pub count: usize,
+}
+
+impl fmt::Display for ReplaceTokenError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "index {} is out of bounds ({})", self.index, self.count)
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for ReplaceTokenError {}
+
+/*
+░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
+╔══════════════════════════════════════════════════════════════════════════════╗
+║                                                                              ║
 ║                                    Tests                                     ║
 ║                                   ¯¯¯¯¯¯¯                                    ║
 ╚══════════════════════════════════════════════════════════════════════════════╝
@@ -1531,4 +1706,12 @@ mod tests {
             );
         }
     }
+    #[test]
+    fn parse_error_display() {
+        assert_eq!(
+            ParseError::NoLeadingBackslash.to_string(),
+            "json pointer is malformed as it does not start with a backslash ('/')"
+        );
+    }
+
 }
